# dfras-docs/dfras-api-swagger.yaml
openapi: 3.0.3
info:
  title: DFRAS API
  description: |
    Delivery Failure Root Cause Analysis System (DFRAS) API
    
    A comprehensive AI-powered platform for analyzing delivery failures and providing predictive insights for logistics operations.
    
    ## Key Features
    - Multi-source data integration (orders, fleet logs, warehouse records, customer feedback)
    - AI-powered root cause analysis using LLM (all-MiniLM-L6-v2)
    - Automated correlation engine for pattern recognition
    - Intelligent reporting with actionable recommendations
    - Predictive analytics for proactive failure prevention
    
    ## Authentication
    All endpoints require JWT authentication. Use the `/auth/login` endpoint to obtain a token.
    
    ## Sample Data
    The system includes comprehensive sample data from `third-assignment-sample-data-set`:
    - 750+ clients across multiple cities
    - 15,000+ orders with delivery status
    - 50+ warehouses with capacity data
    - 200+ drivers with performance metrics
    - External factors (weather, traffic, events)
    - Customer feedback and ratings
  version: 1.0.0
  contact:
    name: DFRAS Support
    email: support@dfras.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local Development Server
  - url: http://localhost:8002
    description: Analytics Service Direct Access
  - url: http://localhost:8001
    description: Data Service Direct Access
security:
  - bearerAuth: []
paths:
...
  /api/data/orders:
    get:
      tags:
        - Data Management
      summary: Get Orders
      description: Retrieve paginated list of orders
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of orders per page
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: status
          in: query
          description: Filter by order status
          required: false
          schema:
            type: string
            enum: [Delivered, Failed, Pending, In-Transit]
        - name: city
          in: query
          description: Filter by delivery city
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      total:
                        type: integer
                        example: 15000
                      pages:
                        type: integer
                        example: 1500
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Order:
      type: object
      properties:
        order_id:
          type: string
          example: "ORD_001"
        client_id:
          type: string
          example: "CLIENT_001"
        customer_name:
          type: string
          example: "John Doe"
        customer_phone:
          type: string
          example: "+91-9876543210"
        delivery_address_line1:
          type: string
          example: "123 Main Street"
        delivery_address_line2:
          type: string
          example: "Apt 4B"
        city:
          type: string
          example: "Mumbai"
        state:
          type: string
          example: "Maharashtra"
        pincode:
          type: string
          example: "400001"
        amount:
          type: number
          format: float
          example: 2500.0
        status:
          type: string
          enum: [Delivered, Failed, Pending, In-Transit]
          example: "Delivered"
        payment_mode:
          type: string
          example: "Credit Card"
        order_date:
          type: string
          format: date-time
        promised_delivery_date:
          type: string
          format: date-time
        actual_delivery_date:
          type: string
          format: date-time
        failure_reason:
          type: string
          example: null
        created_at:
          type: string
          format: date-time
...